feat: apply generateMetadata on server side
