+++ app/page.tsx
@@ -1,74 +1,11 @@
-"use client"
-import Link from "next/link"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { CalendarView } from "@/components/calendar-view"
-import { CardView } from "@/components/card-view"
-import { FilterBar } from "@/components/filter-bar"
-import { Calendar, List } from "lucide-react"
-import { ToggleBar } from "@/components/toggle-bar"
-import { useCardState } from "@/hooks/cardState"
-import { useEffect, useState } from "react"
-import { Hero } from "@/components/hero"
+import Home from "@/components/home"
+import { EPage, generateMetadata } from "@/lib/metadata"
+import { Metadata } from "next"
 
-export default function Home() {
-  const { isCelsius, setIsCelsius } = useCardState()
-  const [activeTab, setActiveTab] = useState("list")
-  const [isXs, setIsXs] = useState(false)
+export const metadata: Metadata = generateMetadata({
+  params: { pageType: EPage.Home, continent: "" },
+})
 
-  useEffect(() => {
-    // Media query for xs screens (sm: 640px) since Calendar View is not visually appealing on xs screens.
-    const mediaQuery = window.matchMedia("(max-width: 639px)")
-    const handleChange = (e: MediaQueryListEvent | MediaQueryList) => {
-      setIsXs(e.matches)
-    }
-    handleChange(mediaQuery)
-    mediaQuery.addEventListener("change", handleChange)
-    return () => mediaQuery.removeEventListener("change", handleChange)
-  }, [])
-
-  useEffect(() => {
-    if (isXs && activeTab === "calendar") {
-      setActiveTab("list")
-    }
-  }, [isXs, activeTab])
-
-  return (
-    <div className="min-h-screen bg-background">
-      <Hero />
-
-      <FilterBar />
-
-      <Tabs value={activeTab} onValueChange={setActiveTab} className="w-full">
-        <div className="container py-4 flex flex-col sm:flex-row justify-between">
-          <TabsList className="grid w-full max-w-md grid-cols-2 border bg-background">
-            <TabsTrigger
-              value="calendar"
-              className="data-[state=active]:bg-accent pointer-events-none opacity-50 sm:pointer-events-auto sm:opacity-100"
-            >
-              <Calendar className="mr-2 h-4 w-4" /> Calendar View
-            </TabsTrigger>
-            <TabsTrigger value="list" className="data-[state=active]:bg-accent">
-              <List className="mr-2 h-4 w-4" /> Card View
-            </TabsTrigger>
-          </TabsList>
-          <div className="flex items-center space-x-2 mt-4 sm:mt-0">
-            <ToggleBar
-              id="temperature"
-              isChecked={isCelsius}
-              onCheckedChange={(checked) => setIsCelsius(checked)}
-              label={`Showing ${isCelsius ? "°C" : "°F"}`}
-            />
-          </div>
-        </div>
-
-        <TabsContent value="calendar" className="mt-0 bg-background">
-          <CalendarView />
-        </TabsContent>
-
-        <TabsContent value="list" className="mt-0 bg-background">
-          <CardView />
-        </TabsContent>
-      </Tabs>
-    </div>
-  )
+export default function Page() {
+  return <Home />
 }
