+++ app/terms/page.tsx
@@ -1,3 +1,10 @@
+import { EPage, generateMetadata } from "@/lib/metadata"
+import { Metadata } from "next"
+
+export const metadata: Metadata = generateMetadata({
+  params: { pageType: EPage.TermsOfUse },
+})
+
 export default function TermsPage() {
   return (
     <div className="container max-w-4xl py-12 px-4 md:px-8 lg:px-12 rounded-xl shadow-lg mt-8 mb-16">
